mogi/requirements-freeze.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
bioblend==0.13.0
biopython==1.75
bioservices==1.6.0
bleach==3.1.0
boto==2.49.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
colorlog==4.0.2
coverage==5.0
cycler==0.10.0
decorator==4.4.1
deepdiff==4.0.9
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.41.0
django-autocomplete-light==3.5.0
django-bootstrap3==12.0.0
django-bulk-update==2.2.0
django-celery-results==1.1.2
django-filter==2.2.0
django-galaxy==0.0.1
django-gfiles==0.0.1
django-mysql==3.3.0
django-nose==1.4.6
django-rainbowtests==0.6.0
django-rest-framework==0.1.0
django-reversion==3.0.5
django-sb-admin==0.3.7
django-simple-history==2.8.0
django-sslserver==0.22
django-tables2==2.2.1
django-tables2-column-shifter==0.5.2
djangorestframework==3.10.3
docutils==0.15.2
easydev==0.9.38
entrypoints==0.3
et-xmlfile==1.0.1
fs==2.2.1
future==0.18.2
gevent==1.4.0
green==3.0.0
greenlet==0.4.15
grequests==0.4.0
idna==2.8
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
isatools==0.10.3
iso8601==0.1.12
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonpickle==1.2
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
kombu==4.6.7
lxml==4.4.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
mysqlclient==1.4.6
mzml2isa==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nose==1.3.7
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.4
ordered-set==3.1.1
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
plotly==4.3.0
pluggy==0.13.1
progressbar2==3.47.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
pronto==0.12.2
ptyprocess==0.6.0
PubChemPy==1.0.4
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.2
pytest-cov==2.8.1
python-dateutil==2.8.1
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
readme-renderer==24.0
requests==2.22.0
requests-cache==0.5.2
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
scipy==1.3.3
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
suds-jurko==0.6
tablib==0.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.37.0
traitlets==4.3.3
twine==2.0.0
Unidecode==1.1.1
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0

